Refactor Xero class: remove date formatting and query string methods to a trait
